% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svdtools.R
\name{exclude_components}
\alias{exclude_components}
\title{Use Single Value Decomposition to reduce information in a matrix by specific components}
\usage{
exclude_components(mat, exclude = 2:num_components(mat))
}
\arguments{
\item{mat}{A numeric matrix or a data frame coercible into a numeric matrix}

\item{exclude}{A vector of integers specifying the components to exclude, defaults to all but the first}
}
\value{
A matrix or an error if the inputs are illogical
}
\description{
Take a numerical matrix and return a matrix produced by leaving out specified components
}
\examples{

# Uses matrix made from the Linux Penguin: Attribution: Larry Ewing <lewing@isc.tamu.edu>
image(noisymatrix, col=gray.colors(65536))
cleanmatrix<-exclude_components(noisymatrix,50:250)
image(cleanmatrix, col=gray.colors(65536))

}
