% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svdtools.R
\docType{data}
\name{noisymatrix}
\alias{noisymatrix}
\title{A numerical matrix of the Linux logo with noise added}
\format{
A Matrix with 421 rows and 500 columns:
}
\source{
\url{https://isc.tamu.edu/~lewing/linux/}
}
\usage{
noisymatrix
}
\description{
A dataset containing the grayscale image of the linux logo with noise added from Wikimedia Commons Attribution: Larry Ewing <lewing@isc.tamu.edu>
}
\keyword{datasets}
