% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svdtools.R
\name{num_components}
\alias{num_components}
\title{Use return the number of SVD components available in a matrix}
\usage{
num_components(mat)
}
\arguments{
\item{mat}{A numeric matrix or a data frame coercible into a numeric matrix}
}
\value{
An integer number of components contained in the matrix
}
\description{
Take a numerical matrix and return the number of SVD components available to reduce
}
\examples{

# Uses matrix made from the Linux Penguin: Attribution: Larry Ewing <lewing@isc.tamu.edu>
\dontrun{
print(num_components(noisymatrix))
}

}
