% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svdtools.R
\name{plot_explanation}
\alias{plot_explanation}
\title{Plot the variance explained by each component of the SVD}
\usage{
plot_explanation(mat, limit = NULL)
}
\arguments{
\item{mat}{A numeric matrix or a data frame coercible into a numeric matrix}

\item{limit}{(Optional) A numeric value specifying the maximum percentage to display.}
}
\value{
A double-plot of the
}
\description{
Take a numerical matrix and plot the variance explained by the components
}
\examples{

# Uses matrix made from the Linux Penguin: Attribution: Larry Ewing <lewing@isc.tamu.edu>
image(noisymatrix, col=gray.colors(65536))
cleanmatrix<-exclude_components(noisymatrix,50:250)
image(cleanmatrix, col=gray.colors(65536))

}
