% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svdtools.R
\name{reduce_components}
\alias{reduce_components}
\title{Use Single Value Decomposition to reduce information in a matrix by components}
\usage{
reduce_components(mat, n = 1)
}
\arguments{
\item{mat}{A numeric matrix or a data frame coercible into a numeric matrix}

\item{n}{An integer number of components to keep. Default is 1}
}
\value{
A matrix or an error if the inputs are illogical
}
\description{
Take a numerical matrix and return a matrix produced only from the number of components specified
}
\examples{

# Uses matrix made from the Linux Penguin: Attribution: Larry Ewing <lewing@isc.tamu.edu>
image(noisymatrix, col=gray.colors(65536))
cleanmatrix<-reduce_components(noisymatrix,50)
image(cleanmatrix, col=gray.colors(65536))

}
