% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svdtools.R
\name{reduce_percentage}
\alias{reduce_percentage}
\title{Use Single Value Decomposition to reduce information in a matrix by percentage}
\usage{
reduce_percentage(mat, p = 0.9)
}
\arguments{
\item{mat}{A numeric matrix or a data frame coercible into a numeric matrix}

\item{p}{Numeric value such that 0 < p <= 1. Defaults to .9}
}
\value{
A matrix or an error if the inputs are illogical
}
\description{
Take a numerical matrix and return a matrix produced only from the percentage of variance explained
}
\examples{

# Uses matrix made from the Linux Penguin: Attribution: Larry Ewing <lewing@isc.tamu.edu>
image(noisymatrix, col=gray.colors(65536))
cleanmatrix<-reduce_percentage(noisymatrix,.988)
image(cleanmatrix, col=gray.colors(65536))

}
