% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svdtools.R
\name{split_matrix}
\alias{split_matrix}
\title{Use Single Value Decomposition to split a matrix into two}
\usage{
split_matrix(mat, exclude)
}
\arguments{
\item{mat}{A numeric matrix or a data frame coercible into a numeric matrix}

\item{exclude}{A vector of integers specifying the components to exclude}
}
\value{
A named list with two matrices - remaining (inverse of excluded) and excluded (same as exclude_components)
}
\description{
Take a numerical matrix and return two same-sized matrices using exclude_components
}
\examples{


\dontrun{
m<-matrix(rbinom(100,5,.1),ncol=10)
sm<-splitmatrix(m,2:4)
m2<-round(sm$remaining+sm$excluded)
m2<-apply(m2,c(1,2),as.integer)
identical(m,m2) # Should be TRUE
}
}
